fixed invoice page
